+++ ft_printf.h
@@ -24,7 +24,7 @@ int	ft_print_c(char c);
 int	ft_print_s(char *s);
 int	ft_print_p(unsigned long long p);
 int	ft_print_di(int n);
-int	ft_print_u(unsigned long long n, int l);
+int	ft_print_u(unsigned int n, int l);
 int	ft_print_xx(unsigned long long n, char formspec, int l);
 int	ft_printf(const char *formspec, ...);
